requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pytest
httpx
schedule
requests
python-multipart
psycopg2-binary
openai